Refactor import paths for authentication button components
- Updated import paths for SignUpButton in LandingCTA and NavAuthButtons components to use the correct file naming convention.
- This change ensures consistency in component imports across the application.
