+++ components/landingCTA.tsx
@@ -1,7 +1,7 @@
 import { ClerkLoading, ClerkLoaded, SignedIn, SignedOut } from "@clerk/nextjs";
 import { Button } from "./ui/button";
 import { Mic } from "lucide-react";
-import { SignUpButton } from "./auth/authButtons";
+import { SignUpButton } from "./auth/auth-buttons";
 import Link from "next/link";
 
 export function LandingCTA() {
