+++ components/navAuthButtons.tsx
@@ -1,5 +1,5 @@
 import { ClerkLoading, ClerkLoaded, SignedIn, SignedOut } from "@clerk/nextjs";
-import { SignInButton, SignUpButton } from "./auth/authButtons";
+import { SignInButton, SignUpButton } from "./auth/auth-buttons";
 import CustomUserButton from "./auth/customUserButton";
 
 export function NavAuthButtons() {
